requirements.txt
langchain>=0.2.17,<0.3
langchain-community>=0.2.17,<0.3
chromadb>=0.5.4,<0.6
python-dotenv>=1.0.1,<2.0
langchain-text-splitters>=0.2.2,<0.3
langchain-google-genai>=1.0.0,<2.0.0
pypdf>=5.0.0,<6.0.0
python-multipart>=0.0.9,<0.1

